requirements.txt
openai>=1.44.0
python-dotenv>=1.0.1
platformdirs>=4.2.0
